refactor: compose BentoGrid with Grid
