+++ packages/components/scripts/primitive-composition-migrations.ts
@@ -44,7 +44,6 @@ export const allowedGridCounts = new Map<string, number>(
     'access-gate/access-gate.css',
     'action-row/action-row.css',
     'approval-card/approval-card.css',
-    'bento-grid/bento-grid.css',
     'blog-section/blog-section.css',
     'calendar/calendar.css',
     'choice-grid/choice-grid.css',
@@ -75,7 +74,6 @@ export const allowedGridCounts = new Map<string, number>(
   ].map((filePath) => [filePath, 1] as const),
 );
 allowedGridCounts.set('action-row/action-row.css', 4);
-allowedGridCounts.set('bento-grid/bento-grid.css', 2);
 allowedGridCounts.set('blog-section/blog-section.css', 4);
 allowedGridCounts.set('calendar/calendar.css', 2);
 allowedGridCounts.set('data-grid/data-grid.css', 2);
@@ -83,6 +81,7 @@ allowedGridCounts.set('description-list/description-list.css', 4);
 allowedGridCounts.set('feature-section/feature-section.css', 8);
 allowedGridCounts.set('footer/footer.css', 2);
 allowedGridCounts.set('form-section/form-section.css', 7);
+allowedGridCounts.set('grid/grid.css', 2);
 allowedGridCounts.set('kanban-board/kanban-board.css', 5);
 allowedGridCounts.set('logo-cloud/logo-cloud.css', 11);
 // The nested submenu's master/detail column split adds one tracked match
