+++ packages/components/src/components/bento-grid/bento-grid.variables.json
@@ -1,1 +1,1 @@
-[]
+["--cinder-grid-columns"]
