+++ packages/components/src/components/grid/grid.schema.json
@@ -26,6 +26,10 @@
       "type": "string",
       "description": "Column gap override. Wins over `gap` for columns."
     },
+    "collapse": {
+      "type": "boolean",
+      "description": "Enables a narrow-screen fallback where the grid becomes a single column."
+    },
     "minItemWidth": {
       "type": "string",
       "description": "Minimum item width for an intrinsic auto-fill grid. When present, this takes\nprecedence over `columns`."
