+++ packages/components/src/components/grid/grid.schema.ts
@@ -29,6 +29,10 @@ const schema = {
       type: 'string',
       description: 'Column gap override. Wins over `gap` for columns.',
     },
+    collapse: {
+      type: 'boolean',
+      description: 'Enables a narrow-screen fallback where the grid becomes a single column.',
+    },
     minItemWidth: {
       type: 'string',
       description:
