+++ packages/components/src/components/grid/grid.svelte
@@ -17,13 +17,18 @@
 
 <script lang="ts">
   import { classNames } from '../../utilities/class-names.ts';
+  import { useResizeObserver } from '../../utilities/use-resize-observer.svelte.ts';
   import type { GridProps } from './grid.types.ts';
+
+  const COLLAPSE_MAX_WIDTH_REM = 48;
+  const FALLBACK_ROOT_FONT_SIZE_PX = 16;
 
   let {
     columns,
     gap,
     rowGap,
     columnGap,
+    collapse = false,
     minItemWidth,
     as = 'div',
     class: customClassName,
@@ -47,12 +52,56 @@
     if (typeof columns === 'string' && columns.length > 0) return columns;
     return undefined;
   });
+
+  let isNarrow = $state(false);
+  let hasMeasuredWidth = $state(false);
+
+  function getCollapseMaxWidthPx(): number {
+    if (typeof window === 'undefined') return COLLAPSE_MAX_WIDTH_REM * FALLBACK_ROOT_FONT_SIZE_PX;
+
+    const rootFontSize = Number.parseFloat(getComputedStyle(document.documentElement).fontSize);
+    const baseFontSize =
+      Number.isFinite(rootFontSize) && rootFontSize > 0 ? rootFontSize : FALLBACK_ROOT_FONT_SIZE_PX;
+    return COLLAPSE_MAX_WIDTH_REM * baseFontSize;
+  }
+
+  function updateNarrowState(width: number): void {
+    if (!Number.isFinite(width) || width <= 0) return;
+
+    hasMeasuredWidth = true;
+    isNarrow = width <= getCollapseMaxWidthPx();
+  }
+
+  function getObservedWidth(entry: ResizeObserverEntry): number {
+    const borderBoxSize = Array.isArray(entry.borderBoxSize)
+      ? entry.borderBoxSize[0]
+      : entry.borderBoxSize;
+
+    return borderBoxSize?.inlineSize ?? entry.contentRect.width;
+  }
+
+  const observeResize = useResizeObserver(
+    (entries) => {
+      const entry = entries[0];
+      if (entry) updateNarrowState(getObservedWidth(entry));
+    },
+    { box: 'border-box', enabled: () => collapse },
+  );
+
+  const observeGrid = (node: HTMLElement) => {
+    updateNarrowState(node.getBoundingClientRect().width);
+    return observeResize(node);
+  };
 </script>
 
 <svelte:element
   this={as}
   {...rest}
+  {@attach observeGrid}
   class={classNames('cinder-grid', customClassName)}
+  data-cinder-collapse={collapse ? '' : undefined}
+  data-cinder-narrow={collapse && isNarrow ? '' : undefined}
+  data-cinder-wide={collapse && hasMeasuredWidth && !isNarrow ? '' : undefined}
   style:--cinder-grid-columns={resolvedColumns}
   style:--cinder-grid-row-gap={rowGap ?? gap}
   style:--cinder-grid-column-gap={columnGap ?? gap}
